% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adh_links.R
\name{get_adh_links}
\alias{get_adh_links}
\title{get_adh_links}
\usage{
get_adh_links(customer_id, token)
}
\arguments{
\item{customer_id}{ADH Customer ID e.g. customers/123}

\item{token}{Access token retrived from function google_auth()}
}
\value{
A list of ad data links.
}
\description{
Lists links between the specified customer and other Google advertising entities.
}
\examples{
\dontrun{
library(adsDataHubR)
token <- google_auth(client_id = client_id, client_secret=client_secret)
total_customers <- get_customers(token)
customer_id <- "customers/123456"
get_adh_links(customer_id, token)
}
}
