% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customers.R
\name{get_customers}
\alias{get_customers}
\title{get_customers}
\usage{
get_customers(token)
}
\arguments{
\item{token}{Access token retrived from function google_auth()}
}
\value{
A dataframe of all customers along with IDs to which authenticated user has access
}
\description{
Fetch Ads Data Hub customers to which the current user has access. Reference: <https://developers.google.com/ads-data-hub/reference/rest/v1/customers.adsDataLinks/list>
}
\examples{
\dontrun{
library(adsDataHubR)
total_customers <- get_customers(token)
}
}
