% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query.R
\name{get_query}
\alias{get_query}
\title{get_query}
\usage{
get_query(query_name, token)
}
\arguments{
\item{query_name}{e.g.customers/123/analysisQueries/abcd1234}

\item{token}{Access token retrived from function google_auth()}
}
\value{
A list of query details
}
\description{
Retrieves the requested analysis query. Reference: <https://developers.google.com/ads-data-hub/reference/rest/v1/customers.analysisQueries/get>
}
\examples{
\dontrun{
library(adsDataHubR)
token <- google_auth(client_id = client_id, client_secret=client_secret)
get_query(query_name, token)
}
}
