% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_auth.R
\name{google_auth}
\alias{google_auth}
\title{google_auth}
\usage{
google_auth(client_id, client_secret)
}
\arguments{
\item{client_id}{Client ID}

\item{client_secret}{Client Secret}
}
\value{
Environment token that can be used in all other functions
}
\description{
Function to login and create a token to be used in all the funtions Reference: <https://developers.google.com/ads-data-hub/guides/quickstart-api>
}
\examples{
\dontrun{
token <- google_auth(client_id = client_id, client_secret=client_secret)
}
}
\seealso{
\url{https://developers.google.com/ads-data-hub/guides/quickstart-api}
}
